pyproject.toml
[tool.poetry]
name = "redmage"
version = "0.0.1"
description = "A component based library for building htmx powered applications."
authors = ["Scott Russell <me@scottrussell.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.26.1"
hype-html = "^1.1.3"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pycodestyle = "^2.10.0"
black = "^23.3.0"
isort = "^5.12.0"
httpx = "^0.24.0"
jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
